+++ src/app/account/additional-components/account-flows-tab/account-flows-subtabs/account-flows-activity-subtab/account-flows-activity-subtab.component.spec.ts
@@ -66,7 +66,7 @@ describe("AccountFlowsActivitySubtabComponent", () => {
         component = fixture.componentInstance;
         component.accountName = MOCK_ACCOUNT_NAME;
         component.accountFlowsData = {
-            activeNav: AccountFlowsNav.DATASETS,
+            activeNav: AccountFlowsNav.PROCESSES,
             flowGroup: [FlowStatus.Finished],
             datasetsFiltersMode: ProcessCardFilterMode.RECENT_ACTIVITY,
         };
@@ -222,7 +222,7 @@ describe("AccountFlowsActivitySubtabComponent", () => {
     it("should check to change navigation", () => {
         const navigateToOwnerViewSpy = spyOn(navigationService, "navigateToOwnerView");
         const event = {
-            nextId: AccountFlowsNav.DATASETS,
+            nextId: AccountFlowsNav.PROCESSES,
         } as NgbNavChangeEvent;
         component.onNavChange(event);
         expect(navigateToOwnerViewSpy).toHaveBeenCalledTimes(1);