+++ src/app/account/services/account-flows-filters.service.spec.ts
@@ -72,7 +72,6 @@ describe("AccountFlowsFiltersService", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
-            applyFilters: true,
         });
     });
 
@@ -95,7 +94,6 @@ describe("AccountFlowsFiltersService", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 1,
             isFirstInitialization: false,
-            applyFilters: true,
         });
     });
 
@@ -126,7 +124,6 @@ describe("AccountFlowsFiltersService", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 3,
             isFirstInitialization: false,
-            applyFilters: true,
         });
         const result = service.setFlowProcessFilters(ProcessCardFilterMode.TRIAGE);
         expect(result).toEqual(
@@ -151,7 +148,6 @@ describe("AccountFlowsFiltersService", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
-            applyFilters: true,
         });
         const result = service.setFlowProcessFilters(ProcessCardFilterMode.UPCOMING_SCHEDULED);
         expect(result).toEqual(
@@ -176,7 +172,6 @@ describe("AccountFlowsFiltersService", () => {
             selectedFlowProcessStates: [FlowProcessEffectiveState.Active],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
-            applyFilters: true,
         });
         const result = service.setFlowProcessFilters(ProcessCardFilterMode.UPCOMING_SCHEDULED);
         expect(result).toEqual(
@@ -201,7 +196,6 @@ describe("AccountFlowsFiltersService", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
-            applyFilters: true,
         });
         const result = service.setFlowProcessFilters(ProcessCardFilterMode.PAUSED);
         expect(result).toEqual(
@@ -226,7 +220,6 @@ describe("AccountFlowsFiltersService", () => {
             selectedFlowProcessStates: [],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
-            applyFilters: false,
         });
         const result = service.setFlowProcessFilters(ProcessCardFilterMode.CUSTOM);
         expect(result).toEqual(
@@ -251,7 +244,6 @@ describe("AccountFlowsFiltersService", () => {
             selectedFlowProcessStates: [FlowProcessEffectiveState.Failing, FlowProcessEffectiveState.PausedManual],
             minConsecutiveFailures: 0,
             isFirstInitialization: false,
-            applyFilters: false,
         });
         const result = service.setFlowProcessFilters(ProcessCardFilterMode.CUSTOM);
         expect(result).toEqual(